requirements.txt
streamlit>=1.39.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pandas>=2.1.0
Pillow>=10.0.0
requests>=2.31.0
yfinance>=0.2.54
plotly>=5.18.0
finnhub-python>=2.4.0
openpyxl>=3.1.0
